BitbayApp/requirements.txt
google-cloud
google-cloud-storage
google-cloud-spanner
